refactor(canvas/render): ts support,  separate context functions from RenderMain and Render
