+++ package.json
@@ -24,6 +24,7 @@
   "devDependencies": {
     "@babel/eslint-parser": "^7.21.3",
     "@types/node": "^18.0.0",
+    "@vue/eslint-config-typescript": "^11.0.3",
     "chokidar": "^3.5.3",
     "concurrently": "^8.2.0",
     "cross-env": "^7.0.3",
