+++ packages/canvas/.eslintrc.cjs
@@ -9,7 +9,6 @@
 * A PARTICULAR PURPOSE. SEE THE APPLICABLE LICENSES FOR MORE DETAILS.
 *
 */
-
 module.exports = {
   env: {
     browser: true,
@@ -17,8 +16,10 @@ module.exports = {
     node: true,
     jest: true
   },
-  extends: ['eslint:recommended', 'plugin:vue/vue3-essential'],
+  extends: ['eslint:recommended', 'plugin:vue/vue3-essential', '@vue/eslint-config-typescript'],
   parser: 'vue-eslint-parser',
+  // 忽略 expected 中的内容
+  ignorePatterns: ['type.d.ts'],
   parserOptions: {
     parser: '@babel/eslint-parser',
     ecmaVersion: 'latest',
@@ -26,7 +27,7 @@ module.exports = {
     requireConfigFile: false,
     babelOptions: {
       parserOpts: {
-        plugins: ['jsx']
+        plugins: ['jsx', 'typescript']
       }
     }
   },
@@ -37,6 +38,7 @@ module.exports = {
     'space-before-function-paren': 'off',
     'vue/multi-word-component-names': 'off',
     'no-use-before-define': 'error',
-    'no-unused-vars': ['error', { ignoreRestSiblings: true, varsIgnorePattern: '^_', argsIgnorePattern: '^_' }]
+    'no-unused-vars': ['error', { ignoreRestSiblings: true, varsIgnorePattern: '^_', argsIgnorePattern: '^_' }],
+    'import/no-inner-modules': 'off'
   }
 }
