+++ packages/canvas/render/src/application-function/bridge.ts
@@ -1,0 +1,16 @@
+import { reset } from '../data-utils'
+
+export function useBridge() {
+  const bridge = {}
+  const setBridge = (data, clear = false) => {
+    clear && reset(bridge)
+    Object.assign(bridge, data)
+  }
+
+  const getBridge = () => bridge
+  return {
+    bridge,
+    setBridge,
+    getBridge
+  }
+}
