+++ packages/canvas/render/src/application-function/global-state.ts
@@ -1,0 +1,35 @@
+import { ref, shallowReactive, watchEffect } from 'vue'
+import { reset } from '../data-utils'
+
+const Func = Function
+
+export function useGlobalState() {
+  const globalState = ref([])
+  const getGlobalState = () => {
+    return globalState.value
+  }
+
+  const setGlobalState = (data = []) => {
+    globalState.value = data
+  }
+  const stores = shallowReactive({})
+  watchEffect(() => {
+    reset(stores)
+    globalState.value.forEach(({ id, state = {}, getters = {} }) => {
+      const computedGetters = Object.keys(getters).reduce(
+        (acc, key) => ({
+          ...acc,
+          [key]: new Func('return ' + getters[key].value)().call(acc, state)
+        }),
+        {}
+      )
+      stores[id] = { ...state, ...computedGetters }
+    })
+  })
+  return {
+    globalState,
+    getGlobalState,
+    setGlobalState,
+    stores
+  }
+}
