+++ packages/canvas/render/src/builtin/CanvasBox.vue
@@ -6,7 +6,7 @@
   </component>
 </template>
 
-<script>
+<script lang="ts">
 import CanvasPlaceholder from './CanvasPlaceholder.vue'
 export default {
   components: {
