+++ packages/canvas/render/src/canvas-function/CanvasEmpty.vue
