+++ packages/canvas/render/src/canvas-function/controller.ts
@@ -1,0 +1,7 @@
+const controller: Record<string, any> = {}
+
+export const setController = (controllerData) => {
+  Object.assign(controller, controllerData)
+}
+
+export const getController = () => controller
