+++ packages/canvas/render/src/canvas-function/index.ts
@@ -1,0 +1,4 @@
+export * from './controller'
+export * from './design-mode'
+export * from './global-notify'
+export * from './custom-renderer'
