+++ packages/canvas/render/src/data-utils.ts
@@ -1,0 +1,12 @@
+import { utils as commonUtils } from '@opentiny/tiny-engine-utils'
+export const { parseFunction: generateFunction } = commonUtils
+
+export const reset = (obj) => {
+  Object.keys(obj).forEach((key) => delete obj[key])
+}
+
+// 规避创建function eslint报错
+export const newFn = (...argv) => {
+  const Fn = Function
+  return new Fn(...argv)
+}
