+++ packages/canvas/render/src/lowcode.ts
@@ -13,14 +13,18 @@
 import { getCurrentInstance, nextTick, provide, inject } from 'vue'
 import { I18nInjectionKey } from 'vue-i18n'
 import { api } from './RenderMain'
-import { collectionMethodsMap, generateFn, globalNotify } from './render'
+import { globalNotify } from './canvas-function'
+import { collectionMethodsMap } from './material-function'
+import { generateFn } from './data-function'
 
 export const lowcodeWrap = (props, context) => {
-  const global = {}
+  const global: {
+    [props: string]: any
+  } = {}
   const instance = getCurrentInstance()
   const router = ''
   const route = ''
-  const { t, locale } = inject(I18nInjectionKey).global
+  const { t, locale } = inject(I18nInjectionKey).global as any
   const emit = context.emit
   const ref = (ref) => instance.refs[ref]
