+++ packages/canvas/render/src/material-function/configure.ts
@@ -1,0 +1,4 @@
+export const configure: Record<string, any> = {}
+export const setConfigure = (configureData) => {
+  Object.assign(configure, configureData)
+}
