+++ packages/canvas/render/src/material-function/index.ts
@@ -1,0 +1,4 @@
+export * from './configure'
+export * from './material-getter'
+export * from './support-collection'
+export * from './support-block-slot-data-for-webcomponent'
