+++ packages/canvas/render/src/material-function/support-block-slot-data-for-webcomponent.ts
@@ -1,0 +1,3 @@
+import { reactive } from 'vue'
+
+export const blockSlotDataMap = reactive<Record<string, any>>({})
