+++ packages/canvas/render/src/material-function/support-collection.ts
@@ -1,0 +1,15 @@
+export const collectionMethodsMap = {}
+
+export const generateCollection = (schema) => {
+  if (schema.componentName === 'Collection' && schema.props?.dataSource && schema.children) {
+    schema.children.forEach((item) => {
+      const fetchData = item.props?.fetchData
+      const methodMatch = fetchData?.value?.match(/this\.(.+?)}/)
+      if (fetchData && methodMatch?.[1]) {
+        const methodName = methodMatch[1].trim()
+        // 缓存表格fetchData对应的数据源信息
+        collectionMethodsMap[methodName] = schema.props.dataSource
+      }
+    })
+  }
+}
