+++ packages/canvas/render/src/runner.ts
@@ -11,20 +11,24 @@
  */
 
 import { createApp } from 'vue'
-import { addScript, addStyle, dynamicImportComponents, updateDependencies } from '../../common'
+import { addScript, addStyle, dynamicImportComponents, updateDependencies } from '../../common/index.js'
 import TinyI18nHost, { I18nInjectionKey } from '@opentiny/tiny-engine-common/js/i18n'
 import Main, { api } from './RenderMain'
 import lowcode from './lowcode'
 import { supportUmdBlock } from './supportUmdBlock'
 
-const dispatch = (name, data) => {
+type ITinyI18nHostI18nHost = typeof TinyI18nHost
+interface IExtendsTinyI18nHost extends ITinyI18nHostI18nHost {
+  lowcode: typeof lowcode
+}
+
+const dispatch = (name: string, data: { detail: any }) => {
   window.parent.document.dispatchEvent(new CustomEvent(name, data))
 }
 
 const initRenderContext = () => {
-  dispatch('beforeCanvasReady')
-
-  TinyI18nHost.lowcode = lowcode
+  dispatch('beforeCanvasReady', null)
+  ;(TinyI18nHost as IExtendsTinyI18nHost).lowcode = lowcode
 
   window.TinyLowcodeComponent = {}
   window.TinyComponentLibs = {}
@@ -72,7 +76,7 @@ const create = async (config) => {
   App.mount(document.querySelector('#app'))
 
   new ResizeObserver(() => {
-    dispatch('canvasResize')
+    dispatch('canvasResize', null)
   }).observe(document.body)
 
   App.config.errorHandler = () => {}
