+++ packages/canvas/render/type.d.ts
@@ -1,0 +1,12 @@
+export declare global {
+  interface Window {
+    TinyLowcodeComponent: Record<string, any>
+    TinyComponentLibs: Record<string, any>
+    blocks: Record<string, any>
+    thirdPartyDeps: {
+      styles: Array<any>
+      scripts: Array<any>
+    }
+    TinyGlobalConfig: Record<string, any>
+  }
+}
